Refactoring to remove snapshot beta support
