+++ pkg/k8sclient/core_test.go
@@ -1,6 +1,6 @@
 /*
  *
- * Copyright © 2022-2023 Dell Inc. or its subsidiaries. All Rights Reserved.
+ * Copyright © 2022-2025 Dell Inc. or its subsidiaries. All Rights Reserved.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -21,7 +21,6 @@ import (
 	"testing"
 
 	vs "github.com/kubernetes-csi/external-snapshotter/client/v4/apis/volumesnapshot/v1"
-	snapv1 "github.com/kubernetes-csi/external-snapshotter/client/v4/apis/volumesnapshot/v1beta1"
 	snapclient "github.com/kubernetes-csi/external-snapshotter/client/v4/clientset/versioned"
 	"github.com/stretchr/testify/suite"
 	v1 "k8s.io/api/core/v1"
@@ -159,20 +158,8 @@ func (suite *CoreTestSuite) TestCreateClients() {
 		suite.NotNil(client)
 	})
 
-	suite.Run("snapshot beta client", func() {
-		client, err := suite.kubeClient.CreateSnapshotBetaClient(namespace)
-		suite.NoError(err)
-		suite.NotNil(client)
-	})
-
 	suite.Run("snapshot content ga client", func() {
 		client, err := suite.kubeClient.CreateSnapshotContentGAClient()
-		suite.NoError(err)
-		suite.NotNil(client)
-	})
-
-	suite.Run("snapshot content beta client", func() {
-		client, err := suite.kubeClient.CreateSnapshotContentBetaClient()
 		suite.NoError(err)
 		suite.NotNil(client)
 	})
@@ -190,18 +177,8 @@ func (suite *CoreTestSuite) TestCreateClients() {
 		suite.Error(err)
 	})
 
-	suite.Run("snapshot beta client", func() {
-		_, err := suite.kubeClient.CreateSnapshotBetaClient(namespace)
-		suite.Error(err)
-	})
-
 	suite.Run("snapshot content ga client", func() {
 		_, err := suite.kubeClient.CreateSnapshotContentGAClient()
-		suite.Error(err)
-	})
-
-	suite.Run("snapshot content beta client", func() {
-		_, err := suite.kubeClient.CreateSnapshotContentBetaClient()
 		suite.Error(err)
 	})
 
@@ -392,17 +369,11 @@ func (suite *CoreTestSuite) TestSnapshotClassExists() {
 
 	suite.kubeClient.Config = &rest.Config{}
 	cset, _ := snapclient.NewForConfig(suite.kubeClient.Config)
-	volumeSnapshotClassBeta := &snapv1.VolumeSnapshotClass{
-		ObjectMeta: metav1.ObjectMeta{
-			Name: "test-snapshot-class",
-		},
-	}
 	volumeSnapshotClass := &vs.VolumeSnapshotClass{
 		ObjectMeta: metav1.ObjectMeta{
 			Name: "test-snapshot-class",
 		},
 	}
-	cset.SnapshotV1beta1().VolumeSnapshotClasses().Create(context.Background(), volumeSnapshotClassBeta, metav1.CreateOptions{})
 	cset.SnapshotV1().VolumeSnapshotClasses().Create(context.Background(), volumeSnapshotClass, metav1.CreateOptions{})
 
 	exists, err := suite.kubeClient.SnapshotClassExists("test-snapshot-class")
