+++ pkg/k8sclient/resources/volumesnapshot/snap_test.go
@@ -13,7 +13,7 @@
  * limitations under the License.
  *
  */
-package v1beta1_test
+package volumesnapshot_test
 
 import (
 	"context"
@@ -21,9 +21,9 @@ import (
 	t1 "testing"
 
 	"github.com/dell/cert-csi/pkg/k8sclient"
-	snapv1 "github.com/dell/cert-csi/pkg/k8sclient/resources/volumesnapshot/v1beta1"
-	v1 "github.com/kubernetes-csi/external-snapshotter/client/v4/apis/volumesnapshot/v1beta1"
-	snapshotv1 "github.com/kubernetes-csi/external-snapshotter/client/v4/clientset/versioned/typed/volumesnapshot/v1beta1"
+	snapv1 "github.com/dell/cert-csi/pkg/k8sclient/resources/volumesnapshot"
+	v1 "github.com/kubernetes-csi/external-snapshotter/client/v4/apis/volumesnapshot/v1"
+	snapshotv1 "github.com/kubernetes-csi/external-snapshotter/client/v4/clientset/versioned/typed/volumesnapshot/v1"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/suite"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
